ISPN-3252 Migrated Configuration API in Spring module
